HAPI/HAPI_DEMO/requirements.txt
torch
transformers
flask
requests
numpyto